pyproject.toml
[project]
name = "psycho-test"
version = "0.1.0"
description = "Fictional Character Personality Quiz"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.0",
    "networkx>=3.6.1",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "gensim>=4.3.0",
    "pillow>=12.1.0",
    "ddgs>=9.10.0",
]
